NP-110 refactoring done
