+++ app/uk/gov/hmrc/uknwauthcheckerapistub/tools/OkMessenger.scala
@@ -17,13 +17,11 @@
 package uk.gov.hmrc.uknwauthcheckerapistub.tools
 
 import play.api.libs.json.{JsValue, Json}
-import play.api.mvc.{AnyContent, Request}
 import uk.gov.hmrc.uknwauthcheckerapistub.models.Eori
 
-trait Helper {
-  val validToken = "Bearer PFZBTElEX1RPS0VOPg=="
+class OkMessenger {
 
-  def makeAJsonRes(eoris: Eori): JsValue = {
+  def makeMessage(eoris: Eori): JsValue = {
 
     val eoriResults: String = eoris.eoris
       .map { anEori =>
@@ -46,16 +44,6 @@ trait Helper {
            |  ]
            |}""".stripMargin)
     eoriMap
-  }
-
-  def validateBearerToken(token: Seq[String]): Boolean =
-    if (token.exists(_ != validToken)) false else true
-
-  def hasValidBearerToken(request: Request[AnyContent]): Boolean = {
-    val valid = request.headers.headers.filter(_._1.contains("authorization"))
-
-    if (valid.nonEmpty && validateBearerToken(valid.map(_._2))) { true }
-    else { false }
   }
 
 }
